Enable twitter previews
Twitter doesn't currently support link previews in the same way that
Facebook and Google+ do for example.

Twitter's version of a link preview is called a 'Twitter Card'.
Currently, Twitter Cards will only be displayed in a post if the site
you're linking to has Twitter Card information available on the site
itself which tells Twitter to display a Twitter Card [1].

This patch adds these headers to our template using the internal Hugo
template, which requires us to add the image for the card as the first
item in the images parameter.

Headers can be validated using the Card validator [2].

[1]: https://developer.twitter.com/en/docs/twitter-for-websites/cards/guides/getting-started
[2]: https://cards-dev.twitter.com/validator